# fix(lsp): correct positions in some scenarios