… cli/lsp/text.rs