1. Keeps all the script snapshots in JavaScript and let TS handle it 
    - Should be ~~faster~~ essentially the same speed and use less memory. Also, lets the TS language service handle updating a file or not.
    - This `ts.ScriptSnapshot.fromString` doesn't do the incremental parsing like before though. It seems tsserver uses this function so maybe best to copy tsserver? It removes a lot of complex code so I think it's worth it.
2. This had the side effect of exposing a bug with multi-byte characters.

Will wait for Kitson to get back before merging.