@kitsonk 2022-04-25
Seems logical.

It would be good to get a baseline of the big file edits LSP bench test off of main and this PR to "prove" the performance improvement (or at least to prove no performance regression).  We don't measure memory consumption on the benchmark tests for the LSP, maybe we should find a way?

We should implement the `.fromString()` I suspect in our snapshots as well.