@dsherret 2022-04-25
> It would be good to get a baseline of the big file edits LSP bench test off of main and this PR to "prove" the performance improvement (or at least to prove no performance regression)

@kitsonk good idea! I just tried it out and it seems the same:

Main:

```
-> Start benchmarking lsp
   - Simple Startup/Shutdown
      (10 runs, mean: 508ms)
   - Big Document/Several Edits 
      (5 runs, mean: 1938ms)
   - Find/Replace
      (10 runs, mean: 84ms)
   - Code Lens
      (10 runs, mean: 610ms)
<- End benchmarking lsp
```

This PR:

```
-> Start benchmarking lsp
   - Simple Startup/Shutdown
      (10 runs, mean: 503ms)
   - Big Document/Several Edits 
      (5 runs, mean: 1901ms)
   - Find/Replace
      (10 runs, mean: 81ms)
   - Code Lens
      (10 runs, mean: 608ms)
```

> We don't measure memory consumption on the benchmark tests for the LSP, maybe we should find a way?

Yeah, I think it would be very good for us to do this so we don't get any regressions. I will look into it tomorrow.