@dsherret 2022-05-02
I looked into a way of tracking memory usage, but it seems like a lot of work so I just opened #14467.

> We should implement the `.fromString()` I suspect in our snapshots as well.

@kitsonk could you clarify what you mean here? I believe I removed our snapshots in this PR.