@kitsonk 2022-05-05
> > We should implement the `.fromString()` I suspect in our snapshots as well.
> 
> @kitsonk could you clarify what you mean here? I believe I removed our snapshots in this PR.

Yes, sorry... also I thought `.fromString()` was an instance method, but it was actually a status method of `ts.ScriptSnapshot`, so basically ignore me.